+++ src/bors/handlers/refresh.rs
@@ -215,9 +215,7 @@ mod tests {
     use crate::bors::handlers::refresh::MOCK_TIME;
     use crate::bors::handlers::trybuild::TRY_BUILD_CHECK_RUN_NAME;
     use crate::database::{MergeableState, OctocrabMergeableState};
-    use crate::tests::mocks::{
-        BorsBuilder, GitHubState, default_pr_number, default_repo_name, run_test,
-    };
+    use crate::tests::{BorsBuilder, GitHubState, default_pr_number, default_repo_name, run_test};
     use chrono::Utc;
     use octocrab::params::checks::{CheckRunConclusion, CheckRunStatus};
     use std::future::Future;
