+++ src/bors/handlers/review.rs
@@ -374,7 +374,7 @@ mod tests {
             RollupMode,
             handlers::{TRY_BRANCH_NAME, trybuild::TRY_MERGE_BRANCH_NAME},
         },
-        tests::mocks::{
+        tests::{
             BorsBuilder, Comment, GitHubState, Permissions, User, default_pr_number,
             default_repo_name, run_test,
         },
