+++ src/tests/mocks/pull_request.rs
@@ -5,14 +5,16 @@ use super::{
     repository::GitHubRepository,
     user::GitHubUser,
 };
-use crate::tests::mocks::repository::PullRequest;
+use crate::github::PullRequestNumber;
+use crate::tests::Branch;
 use crate::{bors::PullRequestStatus, github::GithubRepoName};
 use chrono::{DateTime, Utc};
 use octocrab::models::LabelId;
-use octocrab::models::pulls::MergeableState as OctocrabMergeableState;
+use octocrab::models::pulls::{MergeableState as OctocrabMergeableState, MergeableState};
 use parking_lot::Mutex;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
+use std::time::SystemTime;
 use tokio::sync::mpsc::Sender;
 use url::Url;
 use wiremock::{
@@ -22,6 +24,109 @@ use wiremock::{
 
 pub fn default_pr_number() -> u64 {
     1
+}
+
+#[derive(Clone, Debug)]
+pub struct PullRequest {
+    pub number: PullRequestNumber,
+    pub repo: GithubRepoName,
+    pub labels_added_by_bors: Vec<String>,
+    pub labels_removed_by_bors: Vec<String>,
+    pub comment_counter: u64,
+    pub head_sha: String,
+    pub author: User,
+    pub base_branch: Branch,
+    pub mergeable_state: MergeableState,
+    pub status: PullRequestStatus,
+    pub merged_at: Option<DateTime<Utc>>,
+    pub closed_at: Option<DateTime<Utc>>,
+    pub assignees: Vec<User>,
+    pub description: String,
+    pub title: String,
+    pub labels: Vec<String>,
+}
+
+impl PullRequest {
+    pub fn new(repo: GithubRepoName, number: u64, author: User) -> Self {
+        Self {
+            number: PullRequestNumber(number),
+            repo,
+            labels_added_by_bors: Vec::new(),
+            labels_removed_by_bors: Vec::new(),
+            comment_counter: 0,
+            head_sha: format!("pr-{number}-sha"),
+            author,
+            base_branch: Branch::default(),
+            mergeable_state: MergeableState::Clean,
+            status: PullRequestStatus::Open,
+            merged_at: None,
+            closed_at: None,
+            assignees: Vec::new(),
+            description: format!("Description of PR {number}"),
+            title: format!("Title of PR {number}"),
+            labels: Vec::new(),
+        }
+    }
+}
+
+/// Creates a default pull request with number set to
+/// [default_pr_number].
+impl Default for PullRequest {
+    fn default() -> Self {
+        Self::new(
+            default_repo_name(),
+            default_pr_number(),
+            User::default_pr_author(),
+        )
+    }
+}
+
+impl PullRequest {
+    pub fn check_added_labels(&self, labels: &[&str]) {
+        let added_labels = self
+            .labels_added_by_bors
+            .iter()
+            .map(|s| s.as_str())
+            .collect::<Vec<_>>();
+        assert_eq!(&added_labels, labels);
+    }
+
+    pub fn check_removed_labels(&self, labels: &[&str]) {
+        let removed_labels = self
+            .labels_removed_by_bors
+            .iter()
+            .map(|s| s.as_str())
+            .collect::<Vec<_>>();
+        assert_eq!(&removed_labels, labels);
+    }
+
+    pub fn next_comment_id(&mut self) -> u64 {
+        self.comment_counter += 1;
+        self.comment_counter
+    }
+
+    pub fn merge_pr(&mut self) {
+        self.merged_at = Some(SystemTime::now().into());
+        self.status = PullRequestStatus::Merged;
+    }
+
+    pub fn close_pr(&mut self) {
+        self.closed_at = Some(SystemTime::now().into());
+        self.status = PullRequestStatus::Closed;
+    }
+
+    pub fn reopen_pr(&mut self) {
+        self.closed_at = None;
+        self.status = PullRequestStatus::Open;
+    }
+
+    pub fn ready_for_review(&mut self) {
+        self.status = PullRequestStatus::Open;
+    }
+
+    pub fn convert_to_draft(&mut self) {
+        self.status = PullRequestStatus::Draft;
+    }
 }
 
 pub async fn mock_pull_requests(
@@ -283,8 +388,9 @@ struct GitHubBase {
     ref_field: String,
     sha: String,
 }
+
 #[derive(Serialize)]
-pub(super) struct GitHubPullRequestEventPayload {
+pub struct GitHubPullRequestEventPayload {
     action: String,
     pull_request: GitHubPullRequest,
     changes: Option<GitHubPullRequestChanges>,
